UI component file re-org, added root level popover sheet, fetch all desktop notes
